Move CollectionBadge into common/collections
It is collections vocabulary, and detail-view's nav at shared-platform P2 was importing it from questions at P3. Removing those two edges takes module-boundaries from 52 to 50.
