+++ frontend/src/metabase/common/collections/components/CollectionBadge/CollectionBadge.tsx
