+++ frontend/src/metabase/common/collections/components/CollectionBadge/index.ts
@@ -1,0 +1,1 @@
+export { CollectionBadge } from "./CollectionBadge";
