+++ frontend/src/metabase/detail-view/components/Nav/ModelNav.tsx
@@ -6,7 +6,7 @@ import {
   useGetCollectionQuery,
 } from "metabase/api";
 import { Breadcrumb } from "metabase/common/components/Breadcrumb";
-import { CollectionBadge } from "metabase/questions/components/CollectionBadge";
+import { CollectionBadge } from "metabase/common/collections/components/CollectionBadge";
 import { Group, type GroupProps } from "metabase/ui";
 import * as Urls from "metabase/urls";
 import { getQuestionIdFromVirtualTableId } from "metabase-lib/v1/metadata/utils/saved-questions";
