+++ frontend/src/metabase/nav/components/CollectionBreadcrumbs/CollectionBreadcrumbs.tsx
@@ -4,7 +4,7 @@ import { t } from "ttag";
 import { Breadcrumb } from "metabase/common/components/Breadcrumb";
 import { useToggle } from "metabase/common/hooks/use-toggle";
 import { useTranslateContent } from "metabase/content-translation/hooks";
-import { CollectionBadge } from "metabase/questions/components/CollectionBadge";
+import { CollectionBadge } from "metabase/common/collections/components/CollectionBadge";
 import { ActionIcon, Box, Flex, Icon } from "metabase/ui";
 import * as Urls from "metabase/urls";
 import type {
